% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_metrics.R
\name{calBaselineCellMetrics}
\alias{calBaselineCellMetrics}
\title{Calculate the baseline metrics of segmented cell from cell segmentation output}
\usage{
calBaselineCellMetrics(
  spe,
  metrics = c("total_transciprts", "total_genes", "total_cells", "meanExprsPct_cells"),
  rerun = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{metrics}{A character vector indicating the metrics to be calculated}

\item{rerun}{A logical value indicates whether to recalculate existing metrics.}

\item{verbose}{A logical value indicates whether to print out running messages.}
}
\value{
A SpatialExperiment object
}
\description{
Calculate the baseline metrics of segmented cell from cell segmentation output
}
