% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_metrics.R
\name{calBaselineCellShapeMetrics}
\alias{calBaselineCellShapeMetrics}
\title{Calculate the baseline metrics of cell shapes from cell segmentation output}
\usage{
calBaselineCellShapeMetrics(
  spe,
  metrics = c("cell_area", "elongation", "compactness", "eccentricity", "sphericity",
    "solidity", "convexity", "circularity"),
  use_BPPARAM = BiocParallel::SerialParam(),
  rerun = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{metrics}{A character vector indicating the metrics to be calculated}

\item{use_BPPARAM}{a BiocParallelParam instance determining the parallel
back-end to be used during evaluation.}

\item{rerun}{A logical value indicates whether to recalculate existing metrics.}

\item{verbose}{A logical value indicates whether to print out running messages.}
}
\value{
A SpatialExperiment object
}
\description{
Calculate the baseline metrics of cell shapes from cell segmentation output
}
