% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_metrics.R
\name{calExpressionCorrelation}
\alias{calExpressionCorrelation}
\title{Calculate association with the reference}
\usage{
calExpressionCorrelation(
  spe,
  sce_ref,
  ref_celltype = NULL,
  method = c("pearson"),
  spe_exprs_values = "logcounts",
  ref_exprs_values = "mean"
)
}
\arguments{
\item{spe}{A SpatialExperiment object}

\item{sce_ref}{A SingleCellExperiment object for the reference data}

\item{ref_celltype}{A vector with length of the sample indicating
the cell type label.}

\item{method}{method to compute similarity or distance ("pearson" or
"cosine"). Default is set to "pearson".}

\item{spe_exprs_values}{A string indicating which assay of `spe` contains
the expression values. Default is set to "logcounts".}

\item{ref_exprs_values}{A string indicating which assay of `sce_ref` contains
the expression values. Default is set to "mean".}
}
\value{
A SpatialExperiment object
}
\description{
Calculate association with the reference
}
