% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_metrics.R
\name{calNegMarkerVsDist}
\alias{calNegMarkerVsDist}
\title{Calculate the spatial diversity of base line metrics}
\usage{
calNegMarkerVsDist(
  spe,
  celltype,
  nn_celltype_pair,
  nn_neg_markers_list,
  exprs_threshold = 0,
  distance_breaks = c(seq(0, 50, 10), 100),
  exprs_values = "logcounts"
)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{celltype}{A string or a vector indicating the cell type information.
If it is a string, it will retrive the information from colData.}

\item{nn_celltype_pair}{A pair of cell type.}

\item{nn_neg_markers_list}{A list of negative marker corresponding to each cell type in `nn_celltype_pair`.}

\item{exprs_threshold}{A numeric indicating the threshold to be considered as expressed.}

\item{distance_breaks}{A numeric vector of two or more unique cut points for distance between the pair of cell type.}

\item{exprs_values}{A string indicating which assay of `spe` contains the expression values.}
}
\value{
A SpatialExperiment object
}
\description{
Calculate the spatial diversity of base line metrics
}
