% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDiversity_metrics.R
\name{calSpatialMetricsDiversity}
\alias{calSpatialMetricsDiversity}
\title{Calculate the spatial diversity of base line metrics}
\usage{
calSpatialMetricsDiversity(spe, celltype, bin_width = 500)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{celltype}{A string or a vector indicating the cell type information.
If it is a string, it will retrive the information from colData.}

\item{bin_width}{A numeric indicates the width of the bin.}
}
\value{
A SpatialExperiment object
}
\description{
Calculate the spatial diversity of base line metrics
}
