% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSPA.R
\name{CellSPA}
\alias{CellSPA}
\title{A wrapper function to calculate all metrics}
\usage{
CellSPA(
  spe,
  spe_celltype = NULL,
  sce_ref,
  ref_celltype = sce_ref$celltype,
  positive_marker_list = NULL,
  negative_marker_list = NULL,
  nn_celltype_pair = NULL,
  nn_neg_markers_list = NULL,
  exprs_values = "logcounts",
  use_BPPARAM = BiocParallel::SerialParam(),
  bin_width = 500,
  distance_breaks = c(seq(0, 50, 10), 100),
  verbose = TRUE
)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{spe_celltype}{A string or a vector indicating the cell type information.
If it is a string, it will retrive the information from colData.}

\item{sce_ref}{A SingleCellExperiment object for the reference data}

\item{ref_celltype}{A vector with length of the sample indicating
the cell type label.}

\item{positive_marker_list}{A list of positive markers}

\item{negative_marker_list}{A list of negative markers}

\item{nn_celltype_pair}{A pair of cell type.}

\item{nn_neg_markers_list}{A list of negative marker corresponding to each cell type in `nn_celltype_pair`.}

\item{exprs_values}{A string indicating which assay of `spe` contains the expression values.}

\item{use_BPPARAM}{A BiocParallelParam instance determining the parallel
back-end to be used during evaluation.}

\item{bin_width}{A numeric indicates the width of the bin.}

\item{distance_breaks}{A numeric vector of two or more unique cut points for distance between the pair of cell type.}

\item{verbose}{A logical value indicates whether to print out running messages.}
}
\value{
A SpatialExperiment object
}
\description{
A wrapper function to calculate all metrics
}
