% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{generateMarkerList}
\alias{generateMarkerList}
\title{Generating marker list using the reference data}
\usage{
generateMarkerList(
  sce_ref,
  exprs_values = "mean",
  q = 0.9,
  t = 0.25,
  verbose = TRUE,
  type = "positive"
)
}
\arguments{
\item{sce_ref}{A SingleCellExperiment object}

\item{exprs_values}{A string indicating which assay of `sce_ref` contains
the expression values. Default is set to "mean".}

\item{q}{A numeric indicates the quantile cutoff for selecting the marker list.
By default is 0.9.}

\item{t}{A numeric indicates the cutoff of that a marker allows to appear
in maximum proportion of cell types. By default is 0.25.}

\item{verbose}{A logical value indicates whether to print out running messages.}

\item{type}{A string indicates whether to calculate positive or negative markers.}
}
\value{
A list object
}
\description{
Generating marker list using the reference data
}
