% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_metrics.R
\name{generatePolygon}
\alias{generatePolygon}
\title{Generating polygon}
\usage{
generatePolygon(
  spe,
  use_BPPARAM = BiocParallel::SerialParam(),
  rerun = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{use_BPPARAM}{a BiocParallelParam instance determining the parallel
back-end to be used during evaluation.}

\item{rerun}{A logical value indicates whether to recalculate existing metrics.}

\item{verbose}{A logical value indicates whether to print out running messages.}
}
\value{
A SpatialExperiment object
}
\description{
Generating polygon
}
