% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processingRef}
\alias{processingRef}
\title{Preprocessing the the reference scRNA-seq data}
\usage{
processingRef(sce_ref, celltype, subset_row = NULL)
}
\arguments{
\item{sce_ref}{A SingleCellExperiment object}

\item{celltype}{A vector indicating cell type information.}

\item{subset_row}{A vector specifying the subset of features to keep for the
reference data. This can be a character vector of row names,
an integer vector of row indices or a logical vector.}
}
\value{
A SingleCellExperiment object
}
\description{
Preprocessing the the reference scRNA-seq data
}
