% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processingSPE}
\alias{processingSPE}
\title{Preprocessing the the spatial experiment data.}
\usage{
processingSPE(
  spe,
  qc_range = list(total_transciprts = c(20, 2000), total_genes = c(20, Inf)),
  verbose = TRUE,
  seed = 2023
)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{qc_range}{A list vector with two elements indicating
the range of the total transciprts of the cells to keep.}

\item{verbose}{A logical value indicates whether to print out running messages.}

\item{seed}{A numeric indicates the seed to be set}
}
\value{
A SpatialExperiment object
}
\description{
Preprocessing the the spatial experiment data.
}
