% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readBIDCell}
\alias{readBIDCell}
\title{Read BIDCell output}
\usage{
readBIDCell(data_dir, meta_idx = NULL, tiff_path = NULL, method_name = NULL)
}
\arguments{
\item{data_dir}{a character specifying a single output directory as
returned by BIDCell, corresponding to one sample.}

\item{meta_idx}{a numeric vector indicating the indices of the column that
corresponding to the meta data in the output files.}

\item{tiff_path}{a character specifying the path of a tiff file output from BIDCell.}

\item{method_name}{a character specifying the name of the method.}
}
\value{
A SpatialExperiment object
}
\description{
Read BIDCell output
}
\examples{

data_dir <- system.file("extdata/BIDCell_csv_output", package = "CellSPA")
tiff_path <- system.file("extdata/BIDCell_output_subset.tif", package = "CellSPA")
spe <- readBIDCell(data_dir,
                   tiff_path = tiff_path,
                   method_name = "BIDCell")

}
