% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readMatrixMask_per_sample}
\alias{readMatrixMask_per_sample}
\title{Preprocessing the the spatial experiment data for SimpleSeg output (per sample).}
\usage{
readMatrixMask_per_sample(
  cell_mat,
  cell_coord,
  cell_id,
  mask,
  method_name = "SimpleSeg"
)
}
\arguments{
\item{cell_mat}{A gene by cell matrix indicates the segmented cell expression.}

\item{cell_coord}{A matrix with 3 column indicates the x_coord, y_coord and cell_id of each cell in `cell_mat`.}

\item{cell_id}{A vector indicates the cell_id for the columns of `cell_mat`.}

\item{mask}{The mask results from cell segmentiation method}

\item{method_name}{A string indicates the method name of the cell segmentation.}
}
\description{
Preprocessing the the spatial experiment data for SimpleSeg output (per sample).
}
