% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readSimpleSeg}
\alias{readSimpleSeg}
\title{Preprocessing the the spatial experiment data for SimpleSeg output.}
\usage{
readSimpleSeg(
  spe,
  mask,
  img_id = "imageID",
  cell_id = "cell_id",
  use_BPPARAM = BiocParallel::SerialParam(),
  method_name = "SimpleSeg",
  verbose = TRUE
)
}
\arguments{
\item{spe}{A Spatial Experiment object for multiple image ID}

\item{mask}{The mask results from simpleSeg}

\item{img_id}{A string indicates the image ID column in SpatialExperiment object.}

\item{cell_id}{A string indicates the cell ID column in SpatialExperiment object.}

\item{use_BPPARAM}{BiocParallelParam instance.}

\item{method_name}{A string indicates the method name of the cell segmentation.}

\item{verbose}{A logical value indicates whether to print out running messages.}
}
\description{
Preprocessing the the spatial experiment data for SimpleSeg output.
}
