% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readSimpleSeg_per_sample}
\alias{readSimpleSeg_per_sample}
\title{Preprocessing the the spatial experiment data for SimpleSeg output (per sample).}
\usage{
readSimpleSeg_per_sample(
  spe,
  mask,
  method_name = "SimpleSeg",
  cell_id = "cell_id"
)
}
\arguments{
\item{spe}{A Spatial Experiment object for only one image ID.}

\item{mask}{The mask results from simpleSeg.}

\item{method_name}{A string indicates the method name of the cell segmentation.}

\item{cell_id}{A string indicates the cell ID column in SpatialExperiment object.}
}
\description{
Preprocessing the the spatial experiment data for SimpleSeg output (per sample).
}
