% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readTiffOutput}
\alias{readTiffOutput}
\title{Read Tiff output from BIDCell}
\usage{
readTiffOutput(tiff_path)
}
\arguments{
\item{tiff_path}{a character specifying the path of a tiff file.}
}
\value{
A SpatialExperiment object

tiff_path <- system.file("extdata/BIDCell_output_subset.tif", package = "CellSPA")
tiff_res <- readTiffOutput(tiff_path = tiff_path)
}
\description{
Read Tiff output from BIDCell
}
