% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readVizgen}
\alias{readVizgen}
\title{Read Vizgen output}
\usage{
readVizgen(
  data_dir,
  filter_gene_pattern = "Blank",
  tiff_path = NULL,
  method_name = "Vizgen"
)
}
\arguments{
\item{data_dir}{a character specifying a single output directory from Vizgen.
The data_dir should include one cell_by_gene.csv and one cell_metadata.csv.}

\item{filter_gene_pattern}{Pattern of the genes to be filtered out.}

\item{tiff_path}{a character specifying the path of a tiff file output from BIDCell.}

\item{method_name}{a character specifying the name of the method.}
}
\value{
A SpatialExperiment object
}
\description{
Read Vizgen output
}
