% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readXenium}
\alias{readXenium}
\title{Read Xenium output}
\usage{
readXenium(
  data_dir,
  keep_type = "Gene Expression",
  feature_named_by = "Symbol",
  tiff_path = NULL,
  method_name = "10x"
)
}
\arguments{
\item{data_dir}{a character specifying a single output directory as
returned by Space Ranger, corresponding to one 10x Genomics Xenium sample.}

\item{keep_type}{a character vector indicating the type of features to be
included in the object. By default it is set to `"Gene Expression"`.}

\item{feature_named_by}{a character indicating the way that the features are named.
By default it is set to `"Symbol"`.}

\item{tiff_path}{a character specifying the path of a tiff file output from BIDCell.}

\item{method_name}{a character specifying the name of the method.}
}
\value{
A SpatialExperiment object
}
\description{
Read Xenium output
}
