% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spe}
\alias{spe}
\title{An example SpatialExperiment data}
\format{
## `spe`
A SpatialExperiment as example to illustrate CellSPA:
}
\usage{
spe
}
\description{
A subset of 10x Xenium data in SpatialExperiment format
}
\keyword{datasets}
