% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset}
\alias{subset}
\title{Subset Spatial experiment object by column}
\usage{
subset(spe, col_idx)
}
\arguments{
\item{spe}{A Spatial Experiment object}

\item{col_idx}{The column index to keep.}
}
\value{
A SpatialExperiment object
}
\description{
Subset Spatial experiment object by column
}
