% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethods.R
\name{readTiffOutput}
\alias{readTiffOutput}
\title{Read Tiff output from BIDCell}
\usage{
readTiffOutput(tiff_path)
}
\arguments{
\item{tiff_path}{a character specifying the path of a tiff file.}
}
\value{
A SpatialExperiment object
}
\description{
Read Tiff output from BIDCell
}
